requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.45.0
matplotlib>=3.8.0
numpy>=1.26.0
pillow>=10.0.0
safetensors>=0.4.0
pyyaml>=6.0.0
